feat: create shared FormField component for inputs and AuthLayout for auth pages
