+++ src/App.jsx
@@ -8,8 +8,8 @@ import About from "./Components/About/About.jsx";
 import Details from "./Components/Details/Details.jsx";
 import Networks from "./Components/Networks/Networks.jsx";
 import NotFound from "./Components/NotFound/NotFound.jsx";
-import Login from "./Components/Login/Login.jsx";
-import Register from "./Components/Register/Register.jsx";
+import Login from "../src/Components/Auth/Login/Login.jsx";
+import Register from "../src/Components/Auth/Register/Register.jsx";
 import { Navigate, Route, Routes, useNavigate } from "react-router-dom";
 import { useEffect, useState } from "react";
 import TrendingContextProvider from "./Components/Context/Store.js";
